raytracer: implement
For now, this is the simplest possible implementation. For 4 spheres, it runs in 40-80ms depending on the number of reflection rays. The next steps would be to implement meshes and more shapes, which involves figuring out their intersection functions, along with enough optimizations to run in a managable time. Another possible improvement is supersampling, but that's also very costly.