raycaster: run on multiple threads
This is the biggest "optimization" I could make. It reduced the runtime from over 100 to around 15 seconds.

This channel approach comes with significant overhead due to context switching (which takes up ~1 extra second!). I've tried a batching approach, where the canvas is divided into <thread count> parts, but that ends up being way slower, which is probably because the Linux scheduler doesn't handle it ideally.